`ForexMarketRules` are a kind of `MarketRules`
